Remove unnecessary type casting for user input in AddNumbers.cpp

+++ functions/basics/AddNumbers.cpp
@@ -8,10 +8,10 @@ int add(int a, int b){
 int main(){
     int num1, num2;
     cout << "Enter the first number: ";
-    cin >> (int)num1;
+    cin >> num1;
 
     cout << "Enter the second number: ";
-    cin >> (int)num2;
+    cin >> num2;
 
     int sum = add(num1, num2);
     cout << "The sum of " << num1 << " and " << num2 << " is: " << sum << endl;
